Make extensible AST and parser
- Refactor AST definitions into modules.
- Refactor arithmetic parser into reusable component for operators.
