+++ src/main/scala/edu/colorado/plv/cuanto/jsy/arithmetic/Parser.scala
@@ -1,59 +1,27 @@
-package edu.colorado.plv.cuanto.jsy.arithmetic
+package edu.colorado.plv.cuanto.jsy
+package arithmetic
 
-import edu.colorado.plv.cuanto.jsy.arithmetic.ast._
-import edu.colorado.plv.cuanto.parsing.{ParserLike, RichParsers}
-
-import scala.util.parsing.combinator.JavaTokenParsers
-import scala.util.parsing.input.Reader
+import edu.colorado.plv.cuanto.jsy.common.{JsyParserLike, OpParserLike}
 
 /** Parse into the arithmetic sub-language.
   *
   * @author Bor-Yuh Evan Chang
   */
-object Parser extends JavaTokenParsers with RichParsers with ParserLike[Expr] {
-  override def scan(in: Reader[Char]): Input = in
+object Parser extends OpParserLike with JsyParserLike {
   override def start: Parser[Expr] = expr
-
-  /* Specify the syntax. */
 
   /** Parser for expressions ''expr'': [[Expr]].
     *
     * ''expr'' ::=
     *
     */
-  def expr: Parser[Expr] = binary(bop)
+  def expr: Parser[Expr] = binary
 
-  /** Parser for binary expressions.
-    *
-    * ''binary,,1,,'' ::= ''binary,,2,,'' { ''bop,,1,,'' ''binary,,2,,'' }
-    *
-    * ...
-    *
-    * ''binary,,n,,'' ::= ''unary'' { ''bop,,n,,'' ''unary'' }
-    *
-    */
-  def binary(ops: OpPrecedence): Parser[Expr] = {
-    def level(ops: List[(String, Bop)]): Parser[(Expr, Expr) => Expr] = {
-      val op1 :: t = ops
-      (prebinary(op1) /: t) { (acc, op) => acc | prebinary(op) }
-    }
-    (ops :\ unary) { (lops, acc) => acc * level(lops) }
-  }
-
-  /** ''unary'' ::= ''uop'' ''unary'' | ''atom'' */
-  def unary: Parser[Expr] =
-    positioned {
-      uop ~ unary ^^ { case op ~ e => Unary(op, e) }
-    } |
-    atom
-
-  /** ''atom'' ::= ''float'' | '(' ''expr'' ')' */
-  def atom: Parser[Expr] =
+  /** ''atom'' ::= ''float'' */
+  def opatom: Parser[Expr] =
     positioned {
       floatingPointNumber ^^ (s => N(s.toDouble))
-    } |
-    "(" ~> expr <~ ")" |
-    failure("expected an atom")
+    }
 
   /** ''uop'' ::= '-' */
   def uop: Parser[Uop] =
@@ -69,11 +37,4 @@ object Parser extends JavaTokenParsers with RichParsers with ParserLike[Expr] {
     List("*" -> Times, "/" -> Div)
     /* highest */
   )
-
-  type OpPrecedence = List[List[(String,Bop)]]
-
-  def prebinary(opsyn: (String, Bop)): Parser[(Expr, Expr) => Expr] = {
-    val (csyn, asyn) = opsyn
-    withpos(csyn) ^^ { case (pos, _) => (e1: Expr, e2: Expr) => Binary(asyn, e1, e2) setPos pos }
-  }
 }
