+++ src/main/scala/edu/colorado/plv/cuanto/jsy/common/package.scala
@@ -1,0 +1,18 @@
+package edu.colorado.plv.cuanto.jsy
+
+import edu.colorado.plv.cuanto.parsing.{ParserLike, RichParsers}
+
+import scala.util.parsing.combinator.{JavaTokenParsers, Parsers}
+import scala.util.parsing.input.Reader
+
+/**
+  * @author Bor-Yuh Evan Chang
+  */
+package object common {
+
+  /** Trait for a JavaScripty parser. */
+  trait JsyParserLike extends JavaTokenParsers with RichParsers with ParserLike[Expr] {
+    override def scan(in: Reader[Char]): Input = in
+  }
+
+}
