+++ src/test/scala/edu/colorado/plv/cuanto/jsy/arithmetic/InterpreterSpec.scala
@@ -1,4 +1,5 @@
-package edu.colorado.plv.cuanto.jsy.arithmetic
+package edu.colorado.plv.cuanto.jsy
+package arithmetic
 
 import org.scalatest.{FlatSpec, Matchers}
 import org.scalatest.prop.PropertyChecks
@@ -9,7 +10,7 @@ import org.scalatest.prop.PropertyChecks
 class InterpreterSpec extends FlatSpec with Matchers with PropertyChecks {
   import implicits.stringToExpr
   import Interpreter._
-  import ast._
+  import syntax._
 
   val denoteTests = Table(
     "expression" -> "denotation",
